refactor: rename promptenc to waveenc
