+++ wavesrv/pkg/bufferedpipe/bufferedpipe.go
@@ -15,8 +15,8 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/wavetermdev/waveterm/wavesrv/pkg/promptenc"
 	"github.com/wavetermdev/waveterm/wavesrv/pkg/scbase"
+	"github.com/wavetermdev/waveterm/wavesrv/pkg/waveenc"
 )
 
 const (
@@ -54,7 +54,7 @@ func (pipe *BufferedPipe) GetOutputUrl() (string, error) {
 	qvals := make(url.Values)
 	qvals.Set("key", pipe.Key)
 	qvals.Set("nonce", uuid.New().String())
-	hmacStr, err := promptenc.ComputeUrlHmac([]byte(scbase.WaveAuthKey), BufferedPipeGetterUrl, qvals)
+	hmacStr, err := waveenc.ComputeUrlHmac([]byte(scbase.WaveAuthKey), BufferedPipeGetterUrl, qvals)
 	if err != nil {
 		return "", err
 	}
