+++ wavesrv/pkg/waveenc/hmac.go
@@ -1,7 +1,7 @@
 // Copyright 2024, Command Line Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package promptenc
+package waveenc
 
 import (
 	"crypto/hmac"
